test: sections 테스트와 line 테스트 분리
